Fixed content sources bug
